Unify rollback tests
