+++ core/lib/multivm/src/versions/vm_latest/tests/tester/mod.rs
@@ -1,4 +1,3 @@
-pub(crate) use transaction_test_info::{TransactionTestInfo, TxModifier};
 pub(crate) use vm_tester::{get_empty_storage, InMemoryStorageView, VmTester, VmTesterBuilder};
 pub(crate) use zksync_test_account::{Account, DeployContractsTx, TxType};
